infrastructure/requirements.txt

aws-cdk.core
aws-cdk.aws-lambda
aws-cdk.aws-s3
aws-cdk.aws-iam